lib vtls bearssl

+++ lib/vtls/bearssl.c
@@ -327,7 +327,7 @@ static unsigned x509_end_chain(const br_x509_class **ctx)
   struct x509_context *x509 = (struct x509_context *)ctx;
 
   if(!x509->verifypeer) {
-    return br_x509_decoder_last_error(&x509->decoder);
+    return (unsigned)br_x509_decoder_last_error(&x509->decoder);
   }
 
   return x509->minimal.vtable->end_chain(&x509->minimal.vtable);
@@ -529,7 +529,7 @@ static CURLcode bearssl_set_selected_ciphers(struct Curl_easy *data,
     while(*cipher_end && !bearssl_is_separator(*cipher_end))
       ++cipher_end;
 
-    clen = cipher_end - cipher_start;
+    clen = (size_t)(cipher_end - cipher_start);
     cipher = cipher_start;
 
     cipher_start = cipher_end;
@@ -822,7 +822,7 @@ static CURLcode bearssl_run_until(struct Curl_cfilter *cf,
       if(ret <= 0) {
         return result;
       }
-      br_ssl_engine_sendrec_ack(&backend->ctx.eng, ret);
+      br_ssl_engine_sendrec_ack(&backend->ctx.eng, (size_t)ret);
     }
     else if(state & BR_SSL_RECVREC) {
       buf = br_ssl_engine_recvrec_buf(&backend->ctx.eng, &len);
@@ -835,7 +835,7 @@ static CURLcode bearssl_run_until(struct Curl_cfilter *cf,
       if(ret <= 0) {
         return result;
       }
-      br_ssl_engine_recvrec_ack(&backend->ctx.eng, ret);
+      br_ssl_engine_recvrec_ack(&backend->ctx.eng, (size_t)ret);
     }
   }
 }
